Correction in the post method in appointments, and creation of the getOne method
Error correction of the post method in the appointments controller because the data was being saved but in an infinite loop with several of the same ID together, correcting this by preventing the appointments from happening at the same time, removing the beanUtils and renaming gets for communication coherence between the classes. /
Creation of the getOne method by id. /Fix in appointment Repository adding a custom query to search for a record of the appointments entity method findByDateTimeAndPatientAndDoctor. / Naming correction in appointments dto. / In the doctor and patient models, the @JsonBackReference annotation was added to avoid cyclic references when serializing JSON objects.